// src/types/debtSupplier.ts
import { PaginationParams } from './common';

export interface DebtSupplier {
  id: string | null;
  code: string | null;
  documentType: number | null;
  document: string | null;
  transactionId: string | null;
  customerId: string | null;
  supplierId: string | null;
  supplierCode: string | null;
  supplierName: string | null;
  phoneNumber: string | null;
  employeeId: string | null;
  debt: number | null;
  credit: number | null;
  beginDebt: number | null;
  beginCredit: number | null;
  endDebt: number | null;
  endCredit: number | null;
  deadline: Date | null;
  isAutoGenerated: boolean | null;
  supplierType: number | null;
}

export interface SearchDebtSupplierParams extends PaginationParams {
  supplierId: string | null;
  phoneNumber: string | null;
  debtType: number | null;
  searchDateFrom: Date | null;
  searchDateTo: Date | null;
  type: number | null;
  supplierCode: string | null;
  ndt: number | null;
}

export interface ISupplier {
  id: string;
  code: string | null;
  name: string | null;
  address: string | null;
  type: 1 | 2 | null;
  origin: number | null;
  email: string | null;
  phoneNumber: string;
  bankName: string | null;
  bankBrand: string | null;
  bankNumberAccount: string | null;
  bankAccountHolder: string | null;
  status: number | null;
}